Refactoring of entity relations query
